Migrate draft of model classes in the repository
These include the classes from MCO1 fitted to work with MCO2's
requirements. Additions include:
- VendingMachine abstract class defining common behavior for the two
  types of vending machines
- DispenseResult subclass to better simulate and facilitate vending
  machine output
- SpecialVendingMachine subclass for the special vending machine
- SpecialSlot subclass to encode the extra information needed for a
  special vending machine
- ItemOperation abstract class and related subclasses to facilitate the
  simulation of operations on items (e.g. cooking, topping)
- Preset class to represent items made from other items

This facilitates the creation of the succeeding states, views, and
controllers as the types are already provided. Being able to work on
these view and controller classes also verifies the correctness of the
model classes.